requirements-test.txt
flake8
pytest
pytest-django
pytest-pythonpath
tox
watchdog
